% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{AggregateTiles}
\alias{AggregateTiles}
\title{Quantify aggregated genome tiles}
\usage{
AggregateTiles(object, ...)
}
\arguments{
\item{object}{A Seurat object or ChromatinAssay object}

\item{...}{Additional arguments passed to other methods}
}
\description{
Quantifies fragment counts per cell in fixed-size genome bins across the
whole genome, then removes bins with less than a desired minimum number of
counts in the bin, then merges adjacent tiles into a single region.
}
