% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{AlleleFreq}
\alias{AlleleFreq}
\title{Compute allele frequencies per cell}
\usage{
AlleleFreq(object, ...)
}
\arguments{
\item{object}{A Seurat object, Assay, or matrix}

\item{...}{Arguments passed to other methods}

\item{variants}{A character vector of informative variants to keep. For
example, \code{c("627G>A","709G>A","1045G>A","1793G>A")}.}
}
\value{
Returns a \code{\link[SeuratObject]{Seurat}} object with a new assay
containing the allele frequencies for the informative variants.
}
\description{
Collapses allele counts for each strand and normalize by the total number of
counts at each nucleotide position.
}
