% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{Annotation}
\alias{Annotation}
\alias{Annotation<-}
\title{Annotation}
\usage{
Annotation(object, ...)

Annotation(object, ...) <- value
}
\arguments{
\item{...}{Arguments passed to other methods}

\item{value}{A value to set. Can be NULL, to remove the current annotation
information, or a \code{\link[GenomicRanges]{GRanges}} object. If a
\code{GRanges} object is supplied and the genome information is stored in the
assay, the genome of the new annotations must match the genome of the assay.}
}
\value{
Returns a \code{\link[GenomicRanges]{GRanges}} object
if the annotation data is present, otherwise returns NULL
}
\description{
Get the annotation from a ChromatinAssay
}
