% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{as.ChromatinAssay}
\alias{as.ChromatinAssay}
\title{Convert objects to a ChromatinAssay}
\usage{
as.ChromatinAssay(x, ...)
}
\arguments{
\item{x}{An object to convert to class \code{\link{ChromatinAssay}}}

\item{...}{Arguments passed to other methods}
}
\description{
Convert objects to a ChromatinAssay
}
