% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\docType{data}
\name{CCI_CELLTYPES_COLOR_PALETTE}
\alias{CCI_CELLTYPES_COLOR_PALETTE}
\title{Dictionary of color palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
CCI_CELLTYPES_COLOR_PALETTE
}
\description{
Dictionary of color palettes
}
\keyword{datasets}
