% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_variables.R
\docType{data}
\name{CONSENSUS_APPROACHES}
\alias{CONSENSUS_APPROACHES}
\title{Frequently used variables}
\format{
An object of class \code{character} of length 2.
}
\usage{
CONSENSUS_APPROACHES
}
\description{
Frequently used variables
}
\keyword{datasets}
