% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{ConvertMotifID}
\alias{ConvertMotifID}
\title{Convert between motif name and motif ID}
\usage{
ConvertMotifID(object, ...)
}
\arguments{
\item{object}{A Seurat, ChromatinAssay, or Motif object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a character vector with the same length and order as the
input. Any names or IDs that were not found will be stored as \code{NA}.
}
\description{
Converts from motif name to motif ID or vice versa. To convert common names
to IDs, use the \code{name} parameter. To convert IDs to common names, use
the \code{id} parameter.
}
