% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{FindTopFeatures}
\alias{FindTopFeatures}
\title{Find most frequently observed features}
\usage{
FindTopFeatures(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a \code{\link[SeuratObject]{Seurat}} object
}
\description{
Find top features for a given assay based on total number of counts for the
feature. Can specify a minimum cell count, or a lower percentile
bound to determine the set of variable features. Running this function will
store the total counts and percentile rank for each feature in the feature
metadata for the assay. To only compute the feature metadata, without
changing the variable features for the assay, set \code{min.cutoff=NA}.
}
