% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{Footprint}
\alias{Footprint}
\title{Transcription factor footprinting analysis}
\usage{
Footprint(object, ...)
}
\arguments{
\item{object}{A Seurat or ChromatinAssay object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a \code{\link[SeuratObject]{Seurat}} object
}
\description{
Compute the normalized observed/expected Tn5 insertion frequency
for each position surrounding a set of motif instances.
}
