% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/gbm2json.R
\name{gbm2json}
\alias{gbm2json}
\title{gbm2json function}
\usage{
gbm2json(model, n.trees, ...)
}
\arguments{
\item{model}{gbm model}

\item{n.trees}{number of trees to convert to json format}
}
\description{
This function converts gbm tree trained using gbm package into json format
}
\examples{
gbm2json(model, 10)
}
\keyword{gbm,}
\keyword{gbm2json}
\keyword{json,}

