% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_pairs}
\alias{generate_pairs}
\title{Generate pairs,}
\usage{
generate_pairs(
  v,
  collapse = "__",
  remove_self = TRUE,
  return_undirected = TRUE
)
}
\arguments{
\item{v}{vector with labels}

\item{collapse}{delimiter}

\item{remove_self}{remove self-pairs}

\item{return_undirected}{remove directed pairs, i.e. A-B == B-A}
}
\description{
Generate pairs,
}
