% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_variables.R
\name{get_annot_abbrev_dict}
\alias{get_annot_abbrev_dict}
\title{Dictionary of cell type abbreviations}
\usage{
get_annot_abbrev_dict()
}
\description{
Abbreviation for the different cell types to be used in visualization
}
