% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{get_L2_2}
\alias{get_L2_2}
\title{L4 cell type labels}
\usage{
get_L2_2(return_tme_only = FALSE, return_tumor_only = FALSE)
}
\arguments{
\item{return_tme_only}{return only tme cell types}

\item{return_tumor_only}{return only tumor states}
}
\description{
L4 cell type labels
}
