% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pcs}
\alias{get_pcs}
\title{Function for getting get_pcs}
\usage{
get_pcs(seurat_obj, reduc = "pca", var = 0.05)
}
\arguments{
\item{seurat_obj}{seurat object}

\item{reduc}{layer to use for 'reduction' (default: pca)}

\item{var}{Last point where change of \% of variation is more than X\%}
}
\description{
Function for getting get_pcs
}
