% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_variables.R
\name{get_region_grouped_levels}
\alias{get_region_grouped_levels}
\title{List of regions in GBM data}
\usage{
get_region_grouped_levels()
}
\description{
Levels for the Region_grouped variable (ordered)
}
