% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getMetaInfo.R
\name{getMetaInfo}
\alias{getMetaInfo}
\title{getMetaInfo}
\usage{
getMetaInfo(model, owner = "", description = "", ...)
}
\arguments{
\item{model}{gbm model}

\item{owner}{owner who created the tree. By default blank}

\item{description}{short description. By default blank}
}
\description{
Returns metadata list containing various training parameters for gbm trees
}
\examples{
getMetaInfo(model, owner="ragrawal", description="sample tree")
}
\keyword{gbm,}
\keyword{gbm2json}
\keyword{json,}

