% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{GetMotifData}
\alias{GetMotifData}
\title{Retrieve a motif matrix}
\usage{
GetMotifData(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a \code{\link[SeuratObject]{Seurat}} object
}
\description{
Get motif matrix for given assay
}
