% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getTrees.R
\name{getTrees}
\alias{getTrees}
\title{getTrees}
\usage{
getTrees(model, n.trees)
}
\arguments{
\item{model}{gbm model}

\item{n.trees}{tree to return}
}
\description{
Returns specified tree from gbm model
}
\examples{
getTrees(model, 10)
}
\keyword{gbm,}
\keyword{gbm2json}
\keyword{json,}

