% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getVariables.R
\name{getVariables}
\alias{getVariables}
\title{getVariables}
\usage{
getVariables(model)
}
\arguments{
\item{model}{gbm model}
}
\description{
Returns variables used in building the tree
}
\examples{
getVariables(model)
}
\keyword{gbm,}
\keyword{gbm2json}
\keyword{json,}

