% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{IdentifyVariants}
\alias{IdentifyVariants}
\title{Identify mitochondrial variants}
\usage{
IdentifyVariants(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}
}
\description{
Identify mitochondrial variants present in single cells.
}
