% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{InsertionBias}
\alias{InsertionBias}
\title{Compute Tn5 insertion bias}
\usage{
InsertionBias(object, ...)
}
\arguments{
\item{object}{A Seurat or ChromatinAssay object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a Seurat object
}
\description{
Counts the Tn5 insertion frequency for each DNA hexamer.
}
