% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{Links}
\alias{Links}
\alias{Links<-}
\title{Get or set links information}
\usage{
Links(object, ...)

Links(object, ...) <- value
}
\arguments{
\item{...}{Arguments passed to other methods}

\item{value}{A \code{\link[GenomicRanges]{GRanges}} object}
}
\description{
Get or set the genomic link information for a Seurat object or ChromatinAssay
}
