% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_invasive_signature}
\alias{load_invasive_signature}
\title{Load genelist of invasive signature}
\usage{
load_invasive_signature(name)
}
\arguments{
\item{name}{either 'high' or 'low'}
}
\value{
A list containing two vectors. Invasive up genes at index 1. Invasive down genes at index 2
}
\description{
Load invasive signature derived from comparing PT OPC/NPC1 to Tumour bulk OPC/NPC1 cells
}
