% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{LoadGBM}
\alias{LoadGBM}
\title{Load gbm_regional_study.rds}
\usage{
LoadGBM(mounted_gaitigroup_dirpath = "")
}
\arguments{
\item{mounted_gaitigroup_dirpath}{mounted path for Gaitigroup cluster}
}
\value{
Seurat object with filtered cells from Parsebio and 10X Multiome cohorts. Metadata lookup found on project_status.xlsx on OneDrive
}
\description{
Load gbm_regional_study.rds
}
