% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{LoadRegionColours}
\alias{LoadRegionColours}
\title{Load colour pallette for regions}
\usage{
LoadRegionColours()
}
\value{
Named list for region colour pallette
}
\description{
Load colour pallette for regions
}
