% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{Motifs}
\alias{Motifs}
\alias{Motifs<-}
\title{Get or set a motif information}
\usage{
Motifs(object, ...)

Motifs(object, ...) <- value
}
\arguments{
\item{...}{Arguments passed to other methods}

\item{value}{A \code{\link{Motif}} object}
}
\description{
Get or set the Motif object for a Seurat object or ChromatinAssay.
}
