% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot.gbm.tree.R
\name{plot.gbm.tree}
\alias{plot.gbm.tree}
\title{plot.gbm.tree}
\usage{
\method{plot}{gbm.tree}(m, tree = 1)
}
\arguments{
\item{m}{model}

\item{tree}{tree to plot}
}
\description{
Plots GBM tree using graphviz library
}
\examples{
plot.gbm.tree(model, 1)
}
\keyword{gbm,}
\keyword{gbm2json}
\keyword{json,}

