% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{RunChromVAR}
\alias{RunChromVAR}
\title{Run chromVAR}
\usage{
RunChromVAR(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Additional arguments passed to
\code{\link[chromVAR]{getBackgroundPeaks}}}

\item{genome}{A \code{BSgenome} object or string stating the genome build
recognized by \code{getBSgenome}.}

\item{motif.matrix}{A peak x motif matrix. If NULL, pull the peak x motif
matrix from a Motif object stored in the assay.}

\item{verbose}{Display messages}
}
\value{
Returns a \code{\link[SeuratObject]{Seurat}} object with a new assay
}
\description{
Wrapper to run \code{\link[chromVAR]{chromVAR}} on an assay with a motif
object present. Will return a new Seurat assay with the motif activities
(the deviations in chromatin accessibility across the set of regions) as
a new assay.
}
\details{
See the chromVAR documentation for more information:
\url{https://greenleaflab.github.io/chromVAR/index.html}

See the chromVAR paper: \url{https://www.nature.com/articles/nmeth.4401}
}
