% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{RunSVD}
\alias{RunSVD}
\title{Run singular value decomposition}
\usage{
RunSVD(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a \code{\link[SeuratObject]{Seurat}} object
}
\description{
Run partial singular value decomposition using \code{\link[irlba]{irlba}}
}
