% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{RunTFIDF}
\alias{RunTFIDF}
\title{Compute the term-frequency inverse-document-frequency}
\usage{
RunTFIDF(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a \code{\link[SeuratObject]{Seurat}} object
}
\description{
Run term frequency inverse document frequency (TF-IDF) normalization on a
matrix.
}
\details{
Four different TF-IDF methods are implemented. We recommend using method 1
(the default).
}
\references{
\url{https://en.wikipedia.org/wiki/Latent_semantic_analysis#Latent_semantic_indexing}
}
