% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac-generics.R
\name{SetMotifData}
\alias{SetMotifData}
\title{Set motif data}
\usage{
SetMotifData(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a \code{\link[SeuratObject]{Seurat}} object
}
\description{
Set motif matrix for given assay
}
