% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transpose_df}
\alias{transpose_df}
\title{Transpose dataframe}
\usage{
transpose_df(df, set_index = FALSE)
}
\arguments{
\item{df}{dataframe}

\item{set_index}{set 'col1' as index by default FALSE,}
}
\value{
transposed dataframe
}
\description{
Transpose dataframe
}
