% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{write_patient_fgsea_results}
\alias{write_patient_fgsea_results}
\title{Save Patient FGSEA results}
\usage{
write_patient_fgsea_results(fgsea_results, output_dir, file_header, ranks)
}
\arguments{
\item{fgsea_results}{path to fgsea results}

\item{output_dir}{path to output directory for storing results}

\item{file_header}{name to use for final output directory {output_dir}/{file_header} and generated files {file_header}{}.txt}

\item{ranks}{named vector with ranks}
}
\description{
Save Patient FGSEA results
}
