% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_color_palette}
\alias{load_color_palette}
\title{Load color palette}
\usage{
load_color_palette(name = "CellClass_L2")
}
\value{
named vector with colors
}
\description{
Load pre-defined color palette for cell types, regions etc.
}
