% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_invasive_signature.R
\name{load_invasive_signature}
\alias{load_invasive_signature}
\title{Load genelist of invasive signature}
\usage{
load_invasive_signature(type = "high")
}
\arguments{
\item{type}{choose either 'high' or 'low' (default = 'high')}
}
\value{
A list containing two vectors. Invasive up genes at index 1. Invasive down genes at index 2
}
\description{
Load invasive signature derived from comparing PT OPC/NPC1 to Tumour bulk OPC/NPC1 cells.
}
